==> BANCO GALICIA/1_dict_convertion.ipynb <==
# Nativos
import random as rn
import os
import sys
import gc
import json

#calculo
import numpy as np
import pandas as pd
import scipy as sp

#grafico
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
%matplotlib inline
sns.set(style="whitegrid")

#warning ignore future
import warnings
# warnings.simplefilter(action='ignore', category=FutureWarning)
warnings.filterwarnings("ignore")
gc.collect()

BASE_DIR = os.path.dirname(os.getcwd())
if BASE_DIR not in sys.path: sys.path.append(BASE_DIR)

#from utils import *
#from graphs import *

SEED = 29082013
os.environ['PYTHONHASHSEED']=str(SEED)
np.random.seed(SEED)
rn.seed(SEED)

subfolder = "data"
os.listdir(subfolder)

### Configuración

col_user = 'USER_ID'

convertion = pd.read_csv(
    "{}/conversiones.zip".format(subfolder), dtype={col_user: int, 'mes': int}
).drop_duplicates()[['USER_ID','mes']].sort_values(
    [col_user,'mes']
).reset_index(drop=True)

display(convertion.head())
convertion.shape

def get_trimestre(val):
    if val <= 3:
        return 1
    elif val <= 6:
        return 2
    elif val <= 9:
        return 3
    elif val <= 12:
        return 4
    return 0

convertion['trimestre'] = convertion['mes'].apply(get_trimestre)
for col in convertion.columns:
    print(col, convertion[col].unique().shape)

del convertion['mes']
convertion.head()

convertion = convertion.drop_duplicates()[['USER_ID','trimestre']]
convertion.shape, convertion.groupby(by=['USER_ID', 'trimestre']).size().shape

convertion['USER_ID'].unique().shape

dict_conversion = convertion.groupby(
    by=['trimestre']
)['USER_ID'].apply(
    lambda _: list(_)
).to_dict()

for k, v in dict_conversion.items():
    print(k, ' ',len(v), '  ->',str(round(len(v)*100/11767, 2)), '%')

gc.collect()

with open('convertions.json', 'w') as outfile_names:
    json.dump(dict_conversion, outfile_names)
