==> BANCO GALICIA/2.2_group_percent_pageviews.ipynb <==
# Nativos
import random as rn
import os
import sys
import gc

#calculo
import numpy as np
import pandas as pd
import scipy as sp

#grafico
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
%matplotlib inline
sns.set(style="whitegrid")

#warning ignore future
import warnings
# warnings.simplefilter(action='ignore', category=FutureWarning)
warnings.filterwarnings("ignore")
gc.collect()

BASE_DIR = os.path.dirname(os.getcwd())
if BASE_DIR not in sys.path: sys.path.append(BASE_DIR)

from utils import *
#from graphs import *

SEED = 29082013
os.environ['PYTHONHASHSEED']=str(SEED)
np.random.seed(SEED)
rn.seed(SEED)

subfolder = "data"
os.listdir(subfolder)

## PAGINA SIMPLE

col_user = 'USER_ID'
one_hot_1 = 'PAGE'

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_1]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
pageview[one_hot_1] = pageview[one_hot_1].apply(lambda _: 'page_' + str(_.split('_')[-1]))
print(pageview.shape)
gc.collect()

pageview.head()

cruce = pd.crosstab(
    index=pageview[col_user], columns=pageview[one_hot_1]
)
print(cruce.shape)
cruce.head()

cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
cruce = cruce.reset_index()
gc.collect()

cruce.head()


print(cruce.shape)
cruce.head()

del pageview
gc.collect()

treshold = cruce.shape[0]*1/200
treshold

cols_del = []
for col in list(cruce.columns)[1:]:
    
    columna = cruce[col]
    size = len(columna[columna > 0])
    
    if size < treshold:
        cols_del.append(col) 
        
print("len(cols_del): ", len(cols_del))

cruce = cruce.drop(cols_del, axis=1)
print(cruce.shape)

cols_select = list(cruce.columns)
del cruce
gc.collect()
len(cols_select)

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_1]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

pageview.head()

pageview[one_hot_1] = pageview[one_hot_1].apply(lambda _: 'page_' + str(_.split('_')[-1]))
pageview.head()

trimestral = []

for trimestre in pageview['trimestre'].unique():
    print("/"*50, " trimestre: ", trimestre)
    cruce = pd.crosstab(
        pageview[pageview['trimestre'] == trimestre][col_user], 
        pageview[pageview['trimestre'] == trimestre][one_hot_1]
    )
    cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
    cruce = cruce.reset_index()
    
    cols_cruce = list(cruce.columns)
    cols_del = set(cols_cruce) - set(cols_select)
    print("cols_del: ", len(cols_del))
    
    cols_autocompletar = set(cols_select) - set(cols_cruce)
    print("cols_autocompletar: ", len(cols_autocompletar))
    for col in cols_autocompletar:
        cruce[col] = 0
    
    cruce['trimestre'] = trimestre
    
    trimestral.append(cruce[['trimestre'] + cols_select])
    del cols_cruce
    del cruce
    
    pageview = pageview[pageview['trimestre'] != trimestre]
    gc.collect()

pageview = pd.concat(
    trimestral,
    ignore_index=True,
    axis=0
)

pageview.shape

pageview.to_csv('pageview_simple_trimestral_percent.csv', index=False)
pageview.head()

try:
    gc.collect()
    del pageview
    gc.collect()
except: 
    pass

try:
    gc.collect()
    del trimestral
    gc.collect()
except: 
    pass

try:
    gc.collect()
    del cruce
    gc.collect()
except: 
    pass

gc.collect()

## PAGINA

col_user = 'USER_ID'
one_hot_1 = 'PAGE'

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_1]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

pageview.head()

pageview.dtypes

cruce = pd.crosstab(
    index=pageview[col_user], columns=pageview[one_hot_1]
)
gc.collect()
cruce.shape

del pageview

gc.collect()

cruce['total'] = cruce.sum(axis=1)

cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
cruce = cruce.reset_index()

del cruce['total']
gc.collect()
cruce.head()

### Select columns con recurrencia > 0.5%

cols_del = []
for col in list(cruce.columns)[1:]:
    
    columna = cruce[col]
    size = len(columna[columna > 0])
    
    if size < treshold:
        cols_del.append(col) 
        
print("len(cols_del): ", len(cols_del))

cruce = cruce.drop(cols_del, axis=1)
print(cruce.shape)

cols_select = list(cruce.columns)
del cruce
gc.collect()
len(cols_select)

### Denuevo cargamos las visitas

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_1]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

pageview.head()

for trimestre in pageview['trimestre'].unique():
    print(
        trimestre,
        pageview[pageview['trimestre'] == trimestre]['USER_ID'].unique().shape
    )

trimestral = []

for trimestre in pageview['trimestre'].unique():
    print("/"*50, " trimestre: ", trimestre)
    cruce = pd.crosstab(
        pageview[pageview['trimestre'] == trimestre][col_user], 
        pageview[pageview['trimestre'] == trimestre][one_hot_1]
    )
    cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
    cruce = cruce.reset_index()
    
    cols_cruce = list(cruce.columns)
    cols_del = set(cols_cruce) - set(cols_select)
    print("cols_del: ", len(cols_del))
    
    cols_autocompletar = set(cols_select) - set(cols_cruce)
    print("cols_autocompletar: ", len(cols_autocompletar))
    for col in cols_autocompletar:
        cruce[col] = 0
    
    cruce['trimestre'] = trimestre
    
    trimestral.append(cruce[['trimestre'] + cols_select])
    del cols_cruce
    del cruce
    
    pageview = pageview[pageview['trimestre'] != trimestre]
    gc.collect()

try:
    gc.collect()
    del cruce
    gc.collect
except: 
    pass

try:
    gc.collect()
    del pageview
    gc.collect
except: 
    pass

len(trimestral)

pageview = pd.concat(
    trimestral,
    ignore_index=True,
    axis=0
)

pageview.shape

pageview.to_csv('pageview_trimestral_percent.csv', index=False)
pageview.head()

try:
    gc.collect()
    del pageview
    gc.collect
except: 
    pass

try:
    gc.collect()
    del trimestral
    gc.collect
except: 
    pass

## CATEGORIA

one_hot_2 = 'CONTENT_CATEGORY'

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_2]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

pageview.head()

cruce = pd.crosstab(
    index=pageview[col_user], columns=pageview[one_hot_2]
)
cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
cruce = cruce.reset_index()
gc.collect()

del pageview
gc.collect()
cruce.head()

treshold

cols_del = []
for col in list(cruce.columns)[1:]:
    
    columna = cruce[col]
    size = len(columna[columna > 0])
    
    if size < treshold:
        cols_del.append(col) 

len(cols_del)

cruce = cruce.drop(cols_del, axis=1)
print(cruce.shape)

cols_select = list(cruce.columns)
del cruce
gc.collect()
len(cols_select)

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_2]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

pageview.head()

trimestral = []

for trimestre in pageview['trimestre'].unique():
    print("/"*50, " trimestre: ", trimestre)
    cruce = pd.crosstab(
        pageview[pageview['trimestre'] == trimestre][col_user], 
        pageview[pageview['trimestre'] == trimestre][one_hot_2]
    )
    cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
    cruce = cruce.reset_index()
    
    cols_cruce = list(cruce.columns)
    cols_del = set(cols_cruce) - set(cols_select)
    print("cols_del: ", len(cols_del))
    
    cols_autocompletar = set(cols_select) - set(cols_cruce)
    print("cols_autocompletar: ", len(cols_autocompletar))
    for col in cols_autocompletar:
        cruce[col] = 0
    
    cruce['trimestre'] = trimestre
    
    trimestral.append(cruce[['trimestre'] + cols_select])
    del cols_cruce
    del cruce
    
    pageview = pageview[pageview['trimestre'] != trimestre]
    gc.collect()

try:
    gc.collect()
    del cruce
    gc.collect
except: 
    pass

pageview = pd.concat(
    trimestral,
    ignore_index=True,
    axis=0
)

print(pageview.shape)
pageview.head()

pageview.to_csv('content_category_trimestral_percent.csv', index=False)
pageview.head()

try:
    gc.collect()
    del pageview
    gc.collect
except: 
    pass

try:
    gc.collect()
    del trimestal
    gc.collect
except: 
    pass

## SITE ID

one_hot_2 = 'SITE_ID'

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_2]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

cruce = pd.crosstab(
    index=pageview[col_user], columns=pageview[one_hot_2]
)
cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
cruce = cruce.reset_index()
gc.collect()

del pageview
gc.collect()
cruce.head()

treshold

cols_del = []
for col in list(cruce.columns)[1:]:
    
    columna = cruce[col]
    size = len(columna[columna > 0])
    
    if size < treshold:
        cols_del.append(col) 
        
print(len(cols_del))

cruce = cruce.drop(cols_del, axis=1)
print(cruce.shape)

cols_select = list(cruce.columns)
del cruce
gc.collect()
len(cols_select)

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_2]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

trimestral = []

for trimestre in pageview['trimestre'].unique():
    print("/"*50, " trimestre: ", trimestre)
    cruce = pd.crosstab(
        pageview[pageview['trimestre'] == trimestre][col_user], 
        pageview[pageview['trimestre'] == trimestre][one_hot_2]
    )
    cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
    cruce = cruce.reset_index()
    
    cols_cruce = list(cruce.columns)
    cols_del = set(cols_cruce) - set(cols_select)
    print("cols_del: ", len(cols_del))
    
    cols_autocompletar = set(cols_select) - set(cols_cruce)
    print("cols_autocompletar: ", len(cols_autocompletar))
    for col in cols_autocompletar:
        cruce[col] = 0
    
    cruce['trimestre'] = trimestre
    
    trimestral.append(cruce[['trimestre'] + cols_select])
    del cols_cruce
    del cruce
    
    pageview = pageview[pageview['trimestre'] != trimestre]
    gc.collect()

pageview = pd.concat(
    trimestral,
    ignore_index=True,
    axis=0
)

print(pageview.shape)
pageview.to_csv('site_id_trimestral_percent.csv', index=False)

pageview.head()

### BUSQUEDAS

one_hot_2 = 'ON_SITE_SEARCH_TERM'

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_2]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

cruce = pd.crosstab(
    index=pageview[col_user], columns=pageview[one_hot_2]
)
cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
cruce = cruce.reset_index()
gc.collect()

del pageview
gc.collect()
cruce.head()

cols_del = []
print(treshold)


for col in list(cruce.columns)[1:]:
    
    columna = cruce[col]
    size = len(columna[columna > 0])
    
    if size < treshold:
        cols_del.append(col) 
        
print(len(cols_del))

cruce = cruce.drop(cols_del, axis=1)
print(cruce.shape)

cols_select = list(cruce.columns)
del cruce
gc.collect()
len(cols_select)

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_2]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

trimestral = []

for trimestre in pageview['trimestre'].unique():
    print("/"*50, " trimestre: ", trimestre)
    cruce = pd.crosstab(
        pageview[pageview['trimestre'] == trimestre][col_user], 
        pageview[pageview['trimestre'] == trimestre][one_hot_2]
    )
    cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
    cruce = cruce.reset_index()
    
    cols_cruce = list(cruce.columns)
    cols_del = set(cols_cruce) - set(cols_select)
    print("cols_del: ", len(cols_del))
    
    cols_autocompletar = set(cols_select) - set(cols_cruce)
    print("cols_autocompletar: ", len(cols_autocompletar))
    for col in cols_autocompletar:
        cruce[col] = 0
    
    cruce['trimestre'] = trimestre
    
    trimestral.append(cruce[['trimestre'] + cols_select])
    del cols_cruce
    del cruce
    
    pageview = pageview[pageview['trimestre'] != trimestre]
    gc.collect()

pageview = pd.concat(
    trimestral,
    ignore_index=True,
    axis=0
)

print(pageview.shape)
pageview.to_csv('search_trimestral_percent.csv', index=False)

pageview.head()

cols_select

### DAYS

one_hot_2 = 'tipo_dias'

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', 'FEC_EVENT']]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
pageview['day'] = pageview['FEC_EVENT'].dt.day
pageview['tipo_dias'] = pageview['day'].apply(get_day_cut)
print(pageview.shape)
gc.collect()

cruce = pd.crosstab(
    index=pageview[col_user], columns=pageview[one_hot_2]
)
cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
cruce = cruce.reset_index()
gc.collect()

del pageview
gc.collect()
cruce.head()

cols_del = []
print(treshold)


for col in list(cruce.columns)[1:]:
    
    columna = cruce[col]
    size = len(columna[columna > 0])
    
    if size < treshold:
        cols_del.append(col) 
        
print(len(cols_del))

cruce = cruce.drop(cols_del, axis=1)
print(cruce.shape)

cols_select = list(cruce.columns)
del cruce
gc.collect()
len(cols_select)

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', 'FEC_EVENT']]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
pageview['day'] = pageview['FEC_EVENT'].dt.day
pageview['tipo_dias'] = pageview['day'].apply(get_day_cut)
print(pageview.shape)
gc.collect()

trimestral = []

for trimestre in pageview['trimestre'].unique():
    print("/"*50, " trimestre: ", trimestre)
    cruce = pd.crosstab(
        pageview[pageview['trimestre'] == trimestre][col_user], 
        pageview[pageview['trimestre'] == trimestre][one_hot_2]
    )
    cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
    cruce = cruce.reset_index()
    
    cols_cruce = list(cruce.columns)
    cols_del = set(cols_cruce) - set(cols_select)
    print("cols_del: ", len(cols_del))
    
    cols_autocompletar = set(cols_select) - set(cols_cruce)
    print("cols_autocompletar: ", len(cols_autocompletar))
    for col in cols_autocompletar:
        cruce[col] = 0
    
    cruce['trimestre'] = trimestre
    
    trimestral.append(cruce[['trimestre'] + cols_select])
    del cols_cruce
    del cruce
    
    pageview = pageview[pageview['trimestre'] != trimestre]
    gc.collect()

pageview = pd.concat(
    trimestral,
    ignore_index=True,
    axis=0
)

print(pageview.shape)
pageview.to_csv('rango_dias_trimestral_percent.csv', index=False)

pageview.head()

### SCHEDULE

one_hot_2 = 'horario'

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', 'schedule', 'weekday']]


pageview['USER_ID'] = pageview['USER_ID'].astype(int)
pageview[one_hot_2] = pageview['weekday'] + '_' + pageview['schedule']
print(pageview.shape)
gc.collect()

cruce = pd.crosstab(
    index=pageview[col_user], columns=pageview[one_hot_2]
)
cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
cruce = cruce.reset_index()
gc.collect()

del pageview
gc.collect()
cruce.head()

cols_del = []
print(treshold)


for col in list(cruce.columns)[1:]:
    
    columna = cruce[col]
    size = len(columna[columna > 0])
    
    if size < treshold:
        cols_del.append(col) 
        
print(len(cols_del))

cruce = cruce.drop(cols_del, axis=1)
print(cruce.shape)

cols_select = list(cruce.columns)
del cruce
gc.collect()
len(cols_select)

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', 'schedule', 'weekday']]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
pageview[one_hot_2] = pageview['weekday'] + '_' + pageview['schedule']
print(pageview.shape)
gc.collect()

trimestral = []

for trimestre in pageview['trimestre'].unique():
    print("/"*50, " trimestre: ", trimestre)
    cruce = pd.crosstab(
        pageview[pageview['trimestre'] == trimestre][col_user], 
        pageview[pageview['trimestre'] == trimestre][one_hot_2]
    )
    cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
    cruce = cruce.reset_index()
    
    cols_cruce = list(cruce.columns)
    cols_del = set(cols_cruce) - set(cols_select)
    print("cols_del: ", len(cols_del))
    
    cols_autocompletar = set(cols_select) - set(cols_cruce)
    print("cols_autocompletar: ", len(cols_autocompletar))
    for col in cols_autocompletar:
        cruce[col] = 0
    
    cruce['trimestre'] = trimestre
    
    trimestral.append(cruce[['trimestre'] + cols_select])
    del cols_cruce
    del cruce
    
    pageview = pageview[pageview['trimestre'] != trimestre]
    gc.collect()

pageview = pd.concat(
    trimestral,
    ignore_index=True,
    axis=0
)

print(pageview.shape)
pageview.to_csv('schedule_trimestral_percent.csv', index=False)

pageview.head()

try:
    gc.collect()
    del pageview
except:
    pass

try:
    gc.collect()
    del cruce
except:
    pass

try:
    gc.collect()
    del trimestral
except:
    pass

#### MISMO DIA

one_hot_2 = 'schedule'

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre','schedule']]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

cruce = pd.crosstab(
    index=pageview[col_user], columns=pageview[one_hot_2]
)
cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
cruce = cruce.reset_index()
gc.collect()

del pageview
gc.collect()
cruce.head()

cols_del = []
print(treshold)

for col in list(cruce.columns)[1:]:
    columna = cruce[col]
    size = len(columna[columna > 0])
    
    if size < treshold:
        cols_del.append(col) 

print(len(cols_del))

cruce = cruce.drop(cols_del, axis=1)
print(cruce.shape)

cols_select = list(cruce.columns)
del cruce
gc.collect()
len(cols_select)

one_hot_2

pageview = pd.read_csv(
    "{}/pageview_with_previous_2.csv".format(subfolder), parse_dates=["FEC_EVENT"], dtype=str
).sort_values(
    ["FEC_EVENT"]
)[[col_user, 'trimestre', one_hot_2]]

pageview['USER_ID'] = pageview['USER_ID'].astype(int)
print(pageview.shape)
gc.collect()

trimestral = []

for trimestre in pageview['trimestre'].unique():
    print("/"*50, " trimestre: ", trimestre)
    cruce = pd.crosstab(
        pageview[pageview['trimestre'] == trimestre][col_user], 
        pageview[pageview['trimestre'] == trimestre][one_hot_2]
    )
    cruce = cruce.apply(lambda x: x / x.sum(), axis=1)
    cruce = cruce.reset_index()
    
    cols_cruce = list(cruce.columns)
    cols_del = set(cols_cruce) - set(cols_select)
    print("cols_del: ", len(cols_del))
    
    cols_autocompletar = set(cols_select) - set(cols_cruce)
    print("cols_autocompletar: ", len(cols_autocompletar))
    for col in cols_autocompletar:
        cruce[col] = 0
    
    cruce['trimestre'] = trimestre
    
    trimestral.append(cruce[['trimestre'] + cols_select])
    del cols_cruce
    del cruce
    
    pageview = pageview[pageview['trimestre'] != trimestre]
    gc.collect()

pageview = pd.concat(
    trimestral,
    ignore_index=True,
    axis=0
)

print(pageview.shape)
pageview.to_csv('mismo_trimestral_percent.csv', index=False)
pageview.head()

del pageview
del trimestral
