==> BANCO GALICIA/3.0_agreado_mismo_trim.ipynb <==
# Nativos
import random as rn
import os
import sys
import gc
import json

#calculo
import numpy as np
import pandas as pd
import scipy as sp

#grafico
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
%matplotlib inline
sns.set(style="whitegrid")

#warning ignore future
import warnings
# warnings.simplefilter(action='ignore', category=FutureWarning)
warnings.filterwarnings("ignore")
gc.collect()

BASE_DIR = os.path.dirname(os.getcwd())
if BASE_DIR not in sys.path: sys.path.append(BASE_DIR)

from utils import *
#from graphs import *

SEED = 29082013
os.environ['PYTHONHASHSEED']=str(SEED)
np.random.seed(SEED)
rn.seed(SEED)

subfolder = "data"
os.listdir(subfolder)

pageview = pd.read_csv('pageview_trimestral.csv')
print(pageview.shape)

pageview.head()

with open('convertions.json', 'r') as outfile_names:
    dict_conversion = json.load(outfile_names)

len(dict_conversion)

for trim in pageview['trimestre'].unique():
    print(trim, len(dict_conversion[str(trim)]))

for trim in pageview['trimestre'].unique():
    print("/"*50, trim)
    users_trim = set(pageview[pageview['trimestre'] == trim]['USER_ID'])
    
    print(
        trim, 
        len(dict_conversion[str(trim)]),
        len(users_trim & set(dict_conversion[str(trim)]))
    )

pageview['convertion'] = pageview[['trimestre', 'USER_ID']].apply(
    lambda _: _[1] in dict_conversion[str(_[0])], axis=1
).astype(int)

pageview.head()

pageview.tail()

pageview.dtypes.head(2)

### SPEED

spped = pd.read_csv('speed_trimestral.csv')
print(spped.shape)

spped.head()

print(pageview.shape)
pageview = pageview.merge(spped, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

### DEVICE

device = pd.read_csv('device_trimestral.csv')
print(device.shape)

device.head()

print(pageview.shape)
pageview = pageview.merge(device, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

### SITE_ID

site_id = pd.read_csv('site_id_trimestral.csv')
print(site_id.shape)

site_id.head()

print(pageview.shape)
pageview = pageview.merge(site_id, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

### SCHEDULE

schedule = pd.read_csv('schedule_trimestral.csv')
print(schedule.shape)

schedule.head()

print(pageview.shape)
pageview = pageview.merge(schedule, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

### RANGO DÍAS

dias = pd.read_csv('rango_dias_trimestral.csv')
print(dias.shape)

dias.head()

print(pageview.shape)
pageview = pageview.merge(dias, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

### CATEGORIA DE CONTENIDO

category = pd.read_csv('content_category_trimestral.csv')
print(category.shape)

category.head()

print(pageview.shape)
pageview = pageview.merge(category, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

pageview.head()

nulos = pageview.isnull().sum()
nulos[nulos > 0]

pageview = pageview.fillna(0)

### >> Model

from sklearn.linear_model import LassoCV
from sklearn.feature_selection import SelectFromModel
from lightgbm import LGBMClassifier
from sklearn import model_selection
from sklearn.metrics import roc_auc_score
from sklearn.model_selection import StratifiedShuffleSplit, StratifiedKFold, TimeSeriesSplit, RandomizedSearchCV
from sklearn.model_selection import GridSearchCV
from sklearn import model_selection
from sklearn.metrics import roc_auc_score
from sklearn.model_selection import StratifiedShuffleSplit, StratifiedKFold, TimeSeriesSplit
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import TimeSeriesSplit, cross_val_score, GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn.discriminant_analysis import QuadraticDiscriminantAnalysis
from sklearn.feature_selection import VarianceThreshold
from sklearn.decomposition import PCA
from sklearn.cluster import KMeans
from sklearn.neighbors import KNeighborsClassifier

time_split = TimeSeriesSplit(
    n_splits=4
)

#help(LassoCV)

lasso_cv = LassoCV(
    cv=time_split, 
    selection='random',
    random_state=1234, 
    verbose=1,
    normalize=False,
    n_jobs=4
)

sfm = SelectFromModel(lasso_cv)
sfm.fit(
    pageview.drop(['trimestre', 'USER_ID', 'convertion'], axis=1), 
    pageview['convertion']
)

def var_cols_relevants(sfm, X):
    bool_select_var = list(sfm.get_support())
    list_columns = list(X.columns)
    return list({
       list_columns[i]: bool_select_var[i] for i in range(len(list_columns)) if bool_select_var[i]
    }.keys())

list_var_relevants = var_cols_relevants(sfm, pageview.drop(['trimestre', 'USER_ID', 'convertion'], axis=1))
print(len(list_var_relevants))

pageview[list_var_relevants].head()

import lightgbm as lgb
fit_params = {
    "eval_metric" : 'auc', 
    'verbose': 2
}

lgbm_grid = {
    'bagging_fraction': [0.5, 0.7, 0.9],
    'feature_fraction': [0.25, 0.4, 0.5],
    'n_estimators': [150, 450, 750],
    'max_depth': [3, 5, 7],
    'learning_rate': [0.05, 0.1, 0.15],
    'min_data_in_leaf': [100, 500, 1000],
    'extra_trees': [True, False],
    'deterministic': [True, False],
    'tree_learner': ['serial'],
    'boosting': ['gbdt'],
    'lambda_l2': [0],
    'lambda_l1': [0]
}

def apply_lgbm(X_train_part, X_test_part, y_train, users, add=''):
    model_lgbm = lgb.LGBMClassifier(
        seed=SEED, 
        feature_fraction_seed=SEED, 
        drop_seed=SEED
    )
    grid = RandomizedSearchCV(
        model_lgbm, lgbm_grid, cv=time_split, 
        n_jobs=-1, scoring='roc_auc', verbose=1,
        n_iter=100
    )
    grid.fit(X_train_part, y_train, **fit_params)
    print("-> ", grid.best_score_)
    
    predict_test = grid.best_estimator_.predict_proba(X_test_part)[:,1]
    predict_test = pd.DataFrame(
        {'USER_ID':users, 
         'SCORE':predict_test} 
    )
    display(predict_test.head())
    print(predict_test[predict_test['SCORE'] < 0.01].shape)
    print(predict_test[predict_test['SCORE'] > 0.5].shape)
    print(predict_test[predict_test['SCORE'] > 0.99].shape)
    
    predict_test.groupby(by=['USER_ID'])['SCORE'].mean().to_csv('lgbm_time_split_{}_{}_mean.csv'.format(grid.best_score_, add), index=True)
    predict_test.groupby(by=['USER_ID'])['SCORE'].median().to_csv('lgbm_time_split_{}_{}_median.csv'.format(grid.best_score_, add), index=True)

%%time

apply_lgbm(
    pageview[list_var_relevants], 
    pageview[list_var_relevants], 
    pageview['convertion'], 
    pageview['USER_ID'], 
    add='lasso'
)
