==> BANCO GALICIA/4.0_Graficos.ipynb <==
# Nativos
import random as rn
import os
import sys
import gc
import json

#calculo
import numpy as np
import pandas as pd
import scipy as sp

#grafico
import matplotlib.pyplot as plt
import seaborn as sns
from IPython.display import display
%matplotlib inline
sns.set(style="whitegrid")

#warning ignore future
import warnings
# warnings.simplefilter(action='ignore', category=FutureWarning)
warnings.filterwarnings("ignore")
gc.collect()

BASE_DIR = os.path.dirname(os.getcwd())
if BASE_DIR not in sys.path: sys.path.append(BASE_DIR)

from utils import *
#from graphs import *

SEED = 29082013
os.environ['PYTHONHASHSEED']=str(SEED)
np.random.seed(SEED)
rn.seed(SEED)

subfolder = "data"
os.listdir(subfolder)

pageview = pd.read_csv('pageview_trimestral_percent.csv')
print(pageview.shape)

pageview.head()

pageview.dtypes

with open('convertions.json', 'r') as outfile_names:
    dict_conversion = json.load(outfile_names)

len(dict_conversion)

for trim in pageview['trimestre'].unique():
    print(trim, len(dict_conversion[str(trim)]))

for trim in pageview['trimestre'].unique():
    print(
        trim, 
        len(dict_conversion.get(str(trim + 1), []))
    )

pageview['hizo_conversion_este_mes'] = pageview[['trimestre', 'USER_ID']].apply(
    lambda _: _[1] in dict_conversion.get(str(_[0]), []), axis=1
).astype(int)

pageview.head()

pageview['convertion'] = pageview[['trimestre', 'USER_ID']].apply(
    lambda _: _[1] in dict_conversion.get(str(_[0] + 1), []), axis=1
).astype(int)

pageview.head()

pageview['hizo_conversion_trim_pasado'] = pageview[['trimestre', 'USER_ID']].apply(
    lambda _: _[1] in dict_conversion.get(str(_[0] - 1), []), axis=1
).astype(int)

pageview.head()

pageview['convertion'] = pageview['convertion'].fillna(0)
pd.crosstab(
    pageview['convertion'], pageview['trimestre']
)

pageview['hizo_conversion_este_mes'] = pageview['hizo_conversion_este_mes'].fillna(0)
pd.crosstab(
    pageview['hizo_conversion_este_mes'], pageview['trimestre']
)

pageview['hizo_conversion_trim_pasado'] = pageview['hizo_conversion_trim_pasado'].fillna(0)
pd.crosstab(
    pageview['hizo_conversion_trim_pasado'], pageview['trimestre']
)

pageview['2_conversiones_seguidas'] = pageview[['hizo_conversion_este_mes', 'hizo_conversion_trim_pasado']].apply(
    lambda _: _[0] == 1 and _[1] == 1, axis=1
).astype(int)

pd.crosstab(
    pageview['2_conversiones_seguidas']  , pageview['trimestre']
)

pageview.dtypes.head(2)

### SPEED

spped = pd.read_csv('speed_trimestral_percent.csv')
print(spped.shape)

spped.head()

print(pageview.shape)
pageview = pageview.merge(spped, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

del spped

### DEVICE

device = pd.read_csv('device_trimestral_percent.csv')
print(device.shape)

device.head()

print(pageview.shape)
pageview = pageview.merge(device, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

del device

### SITE_ID

site_id = pd.read_csv('site_id_trimestral_percent.csv')
print(site_id.shape)

site_id.head()

print(pageview.shape)
pageview = pageview.merge(site_id, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

del site_id

### SCHEDULE

schedule = pd.read_csv('schedule_trimestral_percent.csv')
print(schedule.shape)

schedule.head()

print(pageview.shape)
pageview = pageview.merge(schedule, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

del schedule

### RANGO DÍAS

dias = pd.read_csv('rango_dias_trimestral_percent.csv')
print(dias.shape)

dias.head()

print(pageview.shape)
pageview = pageview.merge(dias, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

del dias

### CATEGORIA DE CONTENIDO

category = pd.read_csv('content_category_trimestral_percent.csv')
print(category.shape)

category.head()

print(pageview.shape)
pageview = pageview.merge(category, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

del category

### BUSQUEDA

search = pd.read_csv('search_trimestral_percent.csv')
print(search.shape)

search.head()

print(pageview.shape)
pageview = pageview.merge(search, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

del search

### PAGINA SOLA

simple = pd.read_csv('pageview_simple_trimestral_percent.csv')
print(simple.shape)

simple.head()

print(pageview.shape)
pageview = pageview.merge(simple, how='left', on=["trimestre", "USER_ID"])
print(pageview.shape)

del simple

nulos = pageview.isnull().sum()
nulos[nulos > 0]

pageview = pageview.fillna(0)

pageview.head()

historico_conversiones = pageview[['trimestre', 'USER_ID', '2_conversiones_seguidas', 'hizo_conversion_este_mes', 'hizo_conversion_trim_pasado', 'convertion']]
historico_conversiones.head()

historico_conversiones.shape

del pageview['2_conversiones_seguidas']
del pageview['hizo_conversion_este_mes']
del pageview['hizo_conversion_trim_pasado']
del pageview['convertion']

pageview['trimestre'].unique()

pageview['trimestre'] = pageview['trimestre'].astype(int)

data = []

for trim in [2, 3, 4]:
    print(trim, trim-1)
    
    grupo = pageview[pageview['trimestre'].isin([trim, trim-1])].groupby(by=['USER_ID']).mean().reset_index()
    grupo['trimestre'] = trim
    data.append(grupo)
    del grupo

del pageview

pageview = pd.concat(
    data, 
    ignore_index=True,
    axis=0
)
pageview.head()

pageview.shape

pageview.dtypes.head(3)

historico_conversiones.dtypes.head(3)

historico_conversiones.groupby(by=['trimestre']).size()

for trim in historico_conversiones['trimestre'].unique():
    print(trim, historico_conversiones[historico_conversiones['trimestre'] == trim]['USER_ID'].unique().shape)

pageview.groupby(by=['trimestre']).size()

for trim in pageview['trimestre'].unique():
    print(trim, 
          pageview[pageview['trimestre'] == trim]['USER_ID'].unique().shape,
          len(set(pageview[pageview['trimestre'] == trim]['USER_ID'].unique()) & set(historico_conversiones[historico_conversiones['trimestre'] == trim]['USER_ID'].unique()))
         )

#11589 - 11429, 11489 - 11356, 11314 - 11203

null = historico_conversiones.isnull().sum()
null[null > 0]

historico_conversiones = historico_conversiones.fillna(0)

pageview = pageview.merge(
    historico_conversiones, how='left', on=['USER_ID', 'trimestre'])
pageview.head()

nulo = pageview.isnull().sum()
nulo[nulo > 0]

pageview = pageview.fillna(0)
